Minimal changes required to work with Maven3
