+++ dspace-xmlui/dspace-xmlui-api/src/main/java/org/dspace/app/xmlui/aspect/spider/SpiderTrap.java
